Remove BaseForm as it does not make sense now
